# apps/landing/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fluxby API",
    "version": "1.0.0",
    "description": "API for managing financial transactions, categories, budgets, and analytics. This API is intended for developers building custom interfaces.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/fluxby/fluxby/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/data/import": {
      "post": {
        "summary": "Importeer volledige dataset (maakt bestaande data leeg)",
        "tags": [
          "Data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "accounts": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "transactions": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "budgets": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "categoryRules": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "imports": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "addressBook": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "sharedIbans": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "sharedIbanMerchants": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "nameCleanupRules": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "paymentProviderRules": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data geïmporteerd"
          }
        }
      }
    },
    "/api/data/reset": {
      "delete": {
        "summary": "Reset all data and restore demo profile",
        "description": "Deletes ALL data across all profiles, then creates/restores the demo profile with sample data. This is a destructive operation that cannot be undone.",
        "tags": [
          "Data"
        ],
        "responses": {
          "200": {
            "description": "Data reset and demo profile restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "demoProfileId": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Reset failed"
          }
        }
      }
    },
...
  },
...
}